requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
weaviate-client>=4.4.0
python-dotenv==1.0.1
sentence-transformers==3.1.1
numpy==1.26.4
scikit-learn==1.5.2
rank-bm25==0.2.2
pydantic==2.9.2
